REFACTOR change views names
