+++ src/views/CandidateView.vue
