+++ src/views/CandidatesListView.vue
