+++ src/views/CategoriesListView.vue
