+++ src/views/CategoryView.vue
